pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.9",
    "django-allauth>=65.13.1",
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
